feat: Enhance registration logic, restructure UIFactory, and add customizable icon support (#12)
* fix(register): Phone now registers via patches and derives from Registerable + Moved UIFactory to UI namespace + Usage of Custom Icons

* fix: scene name

* fix: scene loading on mono

* fix: app icons load from Mods/ not Userdata/

* fix: namespace for UI

* fix: stop example phoneapp from appearing
