+++ .gitignore
@@ -9,3 +9,4 @@ Directory.Build.props
 
 # Local assembly references
 ScheduleOneAssemblies/
+*.bin
