+++ S1API/S1API.csproj
@@ -52,7 +52,7 @@
             <HintPath>..\ScheduleOneAssemblies\Il2CppAssemblies\UnityEngine.ImageConversionModule.dll</HintPath>
         </Reference>
     </ItemGroup>
-    
+
     <!-- Mono Specific Deps -->
     <ItemGroup Condition="'$(Configuration)' == 'Mono'">
         <Reference Include="Assembly-CSharp">
@@ -90,18 +90,18 @@
             <HintPath>..\ScheduleOneAssemblies\MelonLoader\0Harmony.dll</HintPath>
         </Reference>
         <Reference Include="Newtonsoft.Json">
-          <HintPath>..\ScheduleOneAssemblies\Managed\Newtonsoft.Json.dll</HintPath>
+            <HintPath>..\ScheduleOneAssemblies\Managed\Newtonsoft.Json.dll</HintPath>
         </Reference>
         <PackageReference Include="LavaGang.MelonLoader" Version="0.7.0" />
     </ItemGroup>
-    
+
     <!-- Automated local deployment -->
     <Target Name="PostBuildIl2Cpp" AfterTargets="Build" Condition=" '$(AutomateLocalDeployment)' == 'true' and '$(Configuration)' == 'Il2Cpp' ">
         <Message Text="Deploying S1API to local Il2Cpp plugins folder" Importance="high"/>
         <Copy SourceFiles=".\bin\Il2Cpp\netstandard2.1\S1API.dll" DestinationFiles="$(LocalIl2CppDeploymentPath)\Plugins\S1API\S1API.Il2Cpp.dll"/>
         <Copy SourceFiles=".\bin\Mono\netstandard2.1\S1API.dll" DestinationFiles="$(LocalIl2CppDeploymentPath)\Plugins\S1API\S1API.Mono.dll"/>
     </Target>
-    
+
     <Target Name="PostBuildMono" AfterTargets="Build" Condition=" '$(AutomateLocalDeployment)' == 'true' and '$(Configuration)' == 'Mono' ">
         <Message Text="Deploying S1API to local Mono plugins folder" Importance="high"/>
         <Copy SourceFiles=".\bin\Il2Cpp\netstandard2.1\S1API.dll" DestinationFiles="$(LocalMonoDeploymentPath)\Plugins\S1API\S1API.Il2Cpp.dll"/>
